2000-05-05  Michael Snyder  <msnyder@seadog.cygnus.com>
        * procfs.c (many functions): change int cmd; to long cmd;
        Solaris /proc API calls this parameter a long, and requires it
        for 64-bit hosts (where sizeof(long) != sizeof(int)).  UnixWare
        calls it an int, but on existing UnixWare hosts, int and long
        are the same size.  So long it must be.  If a future UnixWare
        version has problems with this, we'll have to use autoconf.
        * proc-api.c (write_with_trace): use long instead of int.

+++ gdb/proc-api.c
@@ -440,14 +440,14 @@ static off_t lseek_offset;
 int
 write_with_trace (fd, arg, len, file, line)
      int  fd;
-     int *arg;
+     long *arg;
      size_t len;
      char *file;
      int  line;
 {
   int  i;
   int ret;
-  int opcode = arg[0];
+  long opcode = arg[0];
 
   if (procfs_trace)
     {
